Refactor FroniusSolarApiV1 to use IHttpClient for network requests; update getCurrentPower method to return int and adjust fetchData implementation

+++ src/hal/interfaces/i_inverter_api.h
@@ -23,7 +23,7 @@ enum class InverterStatus {
 class IInverterAPI {
 public:
     virtual ~IInverterAPI() = default;
-    virtual Result<float, InverterError> getCurrentPower() = 0;
+    virtual Result<int, InverterError> getCurrentPower() = 0;
     virtual Result<InverterStatus, InverterError> getStatus() = 0;
     virtual const char* getBrandName() const = 0;
 };
